\name{add_wateryear}
\alias{add_wateryear}
\title{Add a water-year day column}
\description{wrapper for get_wateryear() that adds a column named 'WYD' in the second
column of a dataframe produced by get_waterdata() or get_forecast().}
\usage{add_wateryear(d)}
\arguments{
  \item{d}{a dataframe generated by get_forecast() or get_waterdata().}
}

\value{the supplied dataframe, with 'WYD' in second column.}

\author{Michael Koohafkan}





