\name{dln3}
\alias{dln3}
\title{Density function of the 3-parameter lognormal distribution}
\description{wrapper function for the 3-parameter lognormal distribution defined in
package lmomco. Returns NaN vector instead of NULL if parameters are invalid.}
\usage{dln3(x, zeta, mulog, sigmalog)}
\arguments{
  \item{x}{quantiles}
  \item{zeta}{see lmomco::pdfln3}
  \item{mulog}{see lmomco::pdfln3}
  \item{sigmalog}{see lmomco::pdfln3}
}



\author{Who wrote it}





