\name{dlp3}
\alias{dlp3}
\title{Density function of the (log) Pearson Type III distribution}
\description{wrapper function for the (log) Pearson type III distribution defined in
package lmomco. Returns NaN vector instead of NULL if parameters are invalid.}
\usage{dlp3(x, mu, sigma, gamma)}
\arguments{
  \item{x}{quantiles}
  \item{mu}{see lmomco::pdfpe3}
  \item{sigma}{see lmomco::pdfpe3}
  \item{gamma}{see lmomco::pdfpe3}
}



\author{Michael Koohafkan}





