\name{esptraces}
\docType{data}
\alias{esptraces}

\title{Sample ESP trace data}
\description{Sample ESP trace data. Generated by calling get_forecast() on 5/25/2014.}
\usage{
data(esptraces)
}
\format{
\enumerate{
  \item GMT. Date in YYYY-MM-DD format.
  \item WYD. Date in water-year day format, i.e. 2014-10-01 is 2014001.
  \item Remaining columns are flows for individual ESP traces, in cubic feet per second.
}
}
\examples{
data(esptraces)
summary(esptraces)
}
\source{\url{http://cnrfc.noaa.gov}}
\keyword{datasets}
