\name{flowdurr-package}
\alias{flowdurr-package}
\alias{flowdurr}
\docType{package}
\title{What the package does (short line)}
\description{More about what it does (maybe more than one line)}
\details{
\tabular{ll}{Package: \tab flowdurr\cr
Type: \tab Package\cr
Title: \tab What the package does (short line)\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-18\cr
Author: \tab Who wrote it\cr
Maintainer: \tab Who to complain to <yourfault@somewhere.net>\cr
License: \tab What license is it under?\cr}
}
\author{Who wrote it}

\keyword{ package }


