\name{get_excurves}
\alias{get_excurves}
\title{Calculate exceedance curves.}
\description{generate fitted exceedance curves for plotting.}
\usage{get_excurves(parameterdata, distr, probs = c(seq(0.002, 0.019, 
    0.001), seq(0.02, 0.99, 0.01)))}
\arguments{
  \item{parameterdata}{dataframe of parameters as generated by get_parameters().}
  \item{distr}{The distribution corresponding to the parameter estimates. Can be either
log Pearson III ('lp3') or 3-parameter lognormal ('ln3').}
  \item{probs}{the quantiles used to generate the curve.}
}

\value{a dataframe for plotting exceedance curves for each flow duration. Rows
correspond to probability values contained in column 'p.exceedance'. Column 
names are the flow durations.}

\author{Michael Koohafkan}




\examples{
data(workedexample)
get_excurves(usgs.lp3, distr='lp3',probs=seq(0.02, 0.99, 0.01))
}
