\name{get_parameters}
\alias{get_parameters}
\title{Parameter estimates for exceedance curve fitting.}
\description{estimate the parameters of a log Pearson III or 3-parameter lognormal 
distribution.}
\usage{get_parameters(flowdata, distr = c("lp3", "ln3"), ...)}
\arguments{
  \item{flowdata}{flow duration data to be used for fitting, as created by get_peakavg().}
  \item{distr}{distribution to be fitted, either log Pearson III ('lp3') or lognormal ('ln3').}
  \item{\dots}{other parameters to pass to pel-functions. For more information, see 
lmom::pelln3 and lmom::pelpe3.}
}
\details{The function uses the lmom package to compute method of L-moment 
estimates. See lmom package documentation for information on definitions of 
the log Pearson III and 3-parameter lognormal distributions.}
\value{a dataframe of similar construction to one generated by get_peakavg(). 
Column names are flow durations, named rows are the distribution parameters.}

\author{Who wrote it}




\examples{
#
}
