\name{get_peakavg}
\alias{get_peakavg}
\title{Calculate peak flow durations.}
\description{get the peak average flows for many flow records.}
\usage{get_peakavg(d, flowduration = c(1, 3, 7, 15, 30, 60, 90, 120, 
    183), na.rm = TRUE)}
\arguments{
  \item{d}{a dataframe generated by get_forecast() or get_waterdata().}
  \item{flowduration}{a vector of window sizes for calculating moving averages.}
  \item{na.rm}{logical: remove NA values prior to calculating peak flow durations?}
}
\details{the function uses the rollmean() function in package 'zoo' to
compute flow duration timeseries. The maximum value of each averaged
timeseries is the peak flow for that flow duration.}
\value{a dataframe.}

\author{Who wrote it}




\examples{
#  d1 = add_wateryear(get_forecast(as.Date(Sys.time()) ))
#  d2 = split_by_calendaryear(add_wateryear(get_waterdata(startdate='2010-10-01')))
#  get_peakavg(d1, c(1,3,7,15,30,60,90))
#  get_peakavg(d2, c(1,3,7,15), na.rm=TRUE)
}
