\name{get_ppositions}
\alias{get_ppositions}
\title{Cunnane plotting positions.}
\description{calculate plotting positions for empirical cdf.}
\usage{get_ppositions(pafdata, alpha = 0.4)}
\arguments{
  \item{pafdata}{a data frame generated by get_peakavg().}
  \item{alpha}{shape parameter, default is 0.4.}
}

\value{a data frame of size pafdata containing the plotting 
positions of observations for each record.}
\references{Cunnane 1978, J. Hydrol. 37:205-222
Helsel and Hirsch, Statistical Methods in Water Resources, USGS.}
\author{Who wrote it}




\examples{
#  d1 = add_wateryear(get_forecast(as.Date(Sys.time()) ))
#  d2 = split_by_calendaryear(add_wateryear(get_waterdata(startdate='2010-10-01')))
#  p1 = get_peakavg(d1, c(1,3,7,15,30,60,90))
#  p2 = get_peakavg(d2, c(1,3,7,15), na.rm=TRUE)
#  get_ppositions(p1)
#  get_ppositions(p2, 0.4)
}
