\name{pln3}
\alias{pln3}
\title{Distribution function of the 3-parameter lognormal distribution}
\description{wrapper function for the 3-parameter lognormal distribution defined in
package lmomco. Returns NaN vector instead of NULL if parameters are invalid.}
\usage{pln3(q, zeta, mulog, sigmalog)}
\arguments{
  \item{q}{quantiles}
  \item{zeta}{see lmomco::cdfln3}
  \item{mulog}{see lmomco::cdfln3}
  \item{sigmalog}{see lmomco::cdfln3}
}



\author{Who wrote it}





