\name{plot_all_traces}
\alias{plot_all_traces}
\title{Plot ESP traces.}
\description{plot traces in dataset, as well as mean and median flows.}
\usage{plot_all_traces(d)}
\arguments{
  \item{d}{a dataframe generated by get_forecast() or split_by_*year(get_waterdata() ).}
}

\value{a ggplot plot of all traces, the mean flow, and the median flow.}

\author{Who wrote it}




\examples{
#  d1 = add_wateryear(get_forecast(as.Date(Sys.time()) ))
#  d2 = split_by_calendaryear(add_wateryear(get_waterdata()))
#  plot_all_traces(d1)
#  plot_all_traces(d2)
}
