\name{plot_boxcounts}
\alias{plot_boxcounts}
\title{Box plot and density}
\description{Create box plots for multiple flow durations}
\usage{plot_boxcounts(d, flowdurations = c(1, 3), violin = FALSE, ...)}
\arguments{
  \item{d}{flow data generated by get_forecast() or to_*year()}
  \item{flowdurations}{numeric vector of flow durations in days,}
  \item{violin}{Logical: make violin plot instead of box plot?}
  \item{\dots}{additional variables passed to geom_boxplot() or geom_violin() }
}

\value{a ggplot box plot  }

\author{Michael Koohafkan}




\examples{
data(hspftraces)
hspf.wyear = split_by_wateryear(hspftraces)
plot_boxcounts(hspf.wyear, c(1,3,7))
plot_boxcounts(hspf.wyear, 7, violin=TRUE)
}
