\name{plot_curves}
\alias{plot_curves}
\title{Plot empirical and fitted exceedance curves.}
\description{plot the exceedance curves and empirical data.}
\usage{plot_curves(flowdata, ppos, curvedata, distr, units = c("cf", 
    "af"), logscale = TRUE)}
\arguments{
  \item{flowdata}{flow data generated by get_peakavg().}
  \item{ppos}{plotting positions generated by get_ppositions().}
  \item{curvedata}{fitted curves generated by get_excurves().}
  \item{distr}{the distribution used to generate the fitted curves.}
  \item{units}{the units to be used for displaying flow. Options are cubic feet ('cf')
or acre-feet ('af').}
  \item{logscale}{logical: display log 10 scale on y-axis?}
}

\value{a ggplot2 plot.}

\author{Who wrote it}




\examples{
#
}
