\name{plot_diagnostics}
\alias{plot_diagnostics}
\title{Diagnostic plots for flow duration curves}
\description{Create two diagnostic plots for checking distribution fits}
\usage{plot_diagnostics(dp, distr)}
\arguments{
  \item{dp}{a dataframe of parameter values produced by get_parameters()}
  \item{distr}{the distribution of the supplied parameters. Options are log Pearson type 
III ('lp3') or 3-parameter lognormal ('ln3').}
}



\author{Who wrote it}




\examples{
#
}
