\name{plot_histograms}
\alias{plot_histograms}
\title{Plot histogram and empirical and theoretical densities}
\description{Plot histograms of each flow duration and fitted distribution, with the option of 
plotting the empirical density function and the L-moment estimated distribution.}
\usage{plot_histograms(flowdata, parameterdata, distr = c("lp3", "ln3"), 
    estimatedata = NULL, empirical = TRUE)}
\arguments{
  \item{flowdata}{flow data generated by get_peakavg().}
  \item{parameterdata}{parameter data produced by refine_parameters()}
  \item{distr}{theoretical distribution used, either 3-parameter lognormal ('ln3') or log Pearson
type III distribution ('lp3')}
  \item{estimatedata}{if not NULL, also plot theoretical distribution using L-moment parameter estimates 
produced by get_parameters()}
  \item{empirical}{logical: plot the empirical density function?}
}



\author{Who wrote it}





