\name{plot_trace}
\alias{plot_trace}
\title{Plot a single ESP trace.}
\description{plot an individual trace.}
\usage{plot_trace(d, colname)}
\arguments{
  \item{d}{a dataframe generated by get_forecast() or split_by_*year(get_waterdata() ).}
  \item{colname}{the column name of the trace to be plotted.}
}

\value{a ggplot plot.}

\author{Who wrote it}




\examples{
#  d1 = add_wateryear(get_forecast(as.Date(Sys.time()) ))
#  d2 = split_by_calendaryear(add_wateryear(get_waterdata()))
#  plot_trace(d1, tail(names(d1), 1))
#  plot_trace(d2, tail(names(d2), 1))
}
