\name{qln3}
\alias{qln3}
\title{Quantile function of the 3-parameter lognormal distribution}
\description{wrapper function for the 3-parameter lognormal distribution defined in
package lmomco. Returns NaN vector instead of NULL if parameters are invalid.}
\usage{qln3(p, zeta, mulog, sigmalog)}
\arguments{
  \item{p}{
}
  \item{zeta}{see lmomco::qualn3}
  \item{mulog}{see lmomco::qualn3}
  \item{sigmalog}{see lmomco::qualn3}
}



\author{Michael Koohafkan}





