\name{qlp3}
\alias{qlp3}
\title{Quantile function of the (log) Pearson Type III distribution}
\description{wrapper function for the (log) Pearson type III distribution defined in
package lmomco. Returns NaN vector instead of NULL if parameters are invalid.}
\usage{qlp3(p, mu, sigma, gamma)}
\arguments{
  \item{p}{probabilites}
  \item{mu}{see lmomco::quape3}
  \item{sigma}{see lmomco::quape3}
  \item{gamma}{see lmomco::quape3}
}



\author{Who wrote it}





