\name{refine_parameters}
\alias{refine_parameters}
\title{refine parameters}

\usage{refine_parameters(flowdata, paramdata, distr, returnfits = FALSE)}
\arguments{
  \item{flowdata}{
}
  \item{paramdata}{
}
  \item{distr}{
}
  \item{returnfits}{
}
}



\author{Who wrote it}





