\name{replace_zero_flows}
\alias{replace_zero_flows}
\title{Replace zero values in a dataframe.}
\description{replace numeric values of zero with a specified numeric placeholder.}
\usage{replace_zero_flows(d, r)}
\arguments{
  \item{d}{a dataframe generated by get_forecast() or get_waterdata().}
  \item{r}{the value to substitute, of type numeric.}
}

\value{the same dataframe, with all zero values replaced.}

\author{Who wrote it}




\examples{
#  d = data.frame(x=c(1,2,3), y=c(0, 5, 6), z=c(0,0,9))
#  replace_zero_values(d, 0.001)
#  replace_zero_values(d, 10)
#  replace_zero_values(d, NA)
}
