\name{split_by_calendaryear}
\alias{split_by_calendaryear}
\title{Split a timeseries by calendar year.}
\description{splits a timeseries downloaded using the waterData package by calendar year.}
\usage{split_by_calendaryear(d)}
\arguments{
  \item{d}{a dataframe generated by get_waterdata().}
}

\value{a dataframe with columns 'GMT' and wyXXXX, where XXXX is a year in the record.}

\author{Who wrote it}




\examples{
#  d = get_waterdata()
#  split_by_calendaryear(d)
}
