\name{split_by_wateryear}
\alias{split_by_wateryear}
\title{Split a timeseries by water year.}
\description{splits a timeseries downloaded using the waterData package by water year.}
\usage{split_by_wateryear(d)}
\arguments{
  \item{d}{a dataframe generated by get_waterdata() and add_wateryear().}
}

\value{a dataframe with each column containing data from each 
water year in the record. Missing values are NA.}

\author{Who wrote it}




\examples{
#  d = get_waterdata()
#  split_by_wateryear(d)
}
