\name{subset_forecast}
\alias{subset_forecast}
\title{Subset ESP trace data.}
\description{pull time interval and/or specific locations from forecast data.}
\usage{subset_forecast(d, sc = c("GMT", "WYD"), startdate = NULL, enddate = NULL, 
    location = NULL)}
\arguments{
  \item{d}{a dataframe generated by get_forecast.}
  \item{sc}{name or index of column to use for date, e.g. 'GMT' or 'WYD'. 
Default is 'GMT'.}
  \item{startdate}{the start of interval, date format e.g. as.Date('2012-01-15')
or string 'YYYY-MM-DD', or numeric water year date, 
e.g. October 1st, 2013 is 2014001.}
  \item{enddate}{the last entry in interval, same format as startdate.}
  \item{location}{the region identifier, e.g. Arroyo Hondo is 'AHOC1'.}
}

\value{return subset of data.}

\author{Who wrote it}




\examples{
#  d = get_forecast(as.Date(Sys.time()))
#  d = add_wateryear(d)
#  subset_forecast(d, location='AHOC1')
#  subset_forecast(startdate='2014-04-01', enddate='2014-06-30')
#  subset_forecast(sc='WYD', enddate=2014365)
}
