\name{subset_waterdata}
\alias{subset_waterdata}
\title{Subset downloaded USGS flow data.}
\description{subset waterdata by interval.}
\usage{subset_waterdata(d, sc = c("WYD", "GMT"), startdate = NULL, enddate = NULL, 
    na.cols = TRUE)}
\arguments{
  \item{d}{a dataset generated by split_by_wateryear() or split_by_calendaryear().}
  \item{sc}{search column to use for subsetting by date.}
  \item{startdate}{the first day of the interval, either 'MM-DD' string or 3-digit WYD.}
  \item{enddate}{the last day of the interval, same format as startdate.}
  \item{na.cols}{Logical: remove columns with all values NA?}
}

\value{a subset of the original dataframe.}

\author{Who wrote it}




\examples{
#  d = get_waterdata()
#  subset_waterdata(d, startdate='2013-10-01')
}
