\name{workedexample}
\docType{data}
\alias{workedexample}
\alias{usgs.wyear}
\alias{usgs.clean}
\alias{usgs.3df}
\alias{usgs.fd}
\alias{usgs.pp}
\alias{usgs.lp3}
\alias{usgs.lp3.ref} 
\alias{usgs.fit}
\alias{usgs.exc}

\title{Variables from \code{workedexample} demo.}
\description{Collection of variables produced in the \code{workedexample} demo. Used for testing and examples.}
\usage{
data(workedexample)
}
\format{
\enumerate{
  \item usgs.wyear
  \item usgs.clean
  \item usgs.3df
  \item usgs.fd
  \item usgs.pp
  \item usgs.lp3
  \item usgs.lp3.ref
  \item usgs.fit
  \item usgs.exc
}
}
\examples{
data(workedexample)
ls()
}
\keyword{datasets}
