\name{add_wateryear}
\alias{add_wateryear}
\title{Add a wateryear column.}
\description{wrapper for get_wateryear() that adds a column named 'WYD' in the second
column of a dataframe produced by get_waterdata() or get_forecast().}
\usage{add_wateryear(d)}
\arguments{
  \item{d}{a dataframe generated by get_forecast() or get_waterdata().}
}

\value{the supplied dataframe, with 'WYD' in second column.}

\author{Who wrote it}




\examples{
#  d = data.frame(GMT=as.Date(c('2012-10-01', '2013-01-01', '2013-09-30')),
#                 x=c(10,20,30))
#  add_wateryear(d)
}
