\name{flowdurr-package}
\alias{flowdurr-package}
\alias{flowdurr}
\docType{package}
\title{Functions for calculating exceedence curves using CNRFC ESP traces and USGS flow data.}
\description{This package was created by SFPUC to assist with the use of CNRFC ensemble data and USGS streamflow data for performing tasks related reservoir management and forecasting. The package provides functionality for accessing USGS data provided by the waterData package; downloading CNRFC trace data over HTTP; data formatting, subsetting and visualization; computing flow duration curves; calculating plotting positions for empirical cumulative distribution functions; fitting of Pearson type-III and 3-parameter lognormal distributions; and computing quantiles.}
\details{
\tabular{ll}{Package: \tab flowdurr\cr
Type: \tab Package\cr
Title: \tab Functions for calculating exceedence curves using CNRFC ESP traces and USGS flow data.\cr
Version: \tab 0.2\cr
Date: \tab 2014-05-12\cr
Author: \tab Michael Koohafkan\cr
Maintainer: \tab Michael Koohafkan <mkoohafkan@sfwater.org>\cr
Depends: \tab R (>= 3.1.0), waterData\cr
Imports: \tab zoo, fitdistrplus, rgenoud, scales, ggplot2, reshape2, lmomco\cr
Suggests: \tab knitr\cr
LazyData: \tab true\cr
VignetteBuilder: \tab knitr\cr
License: \tab GPL\cr}
}
\author{Michael Koohafkan}

\keyword{ package }


