\name{get_ppositions}
\alias{get_ppositions}
\title{Cunnane plotting positions.}
\description{calculate plotting positions for empirical cdf.}
\usage{get_ppositions(pafdata, alpha = 0.4)}
\arguments{
  \item{pafdata}{a data frame generated by get_peakavg().}
  \item{alpha}{shape parameter, default is 0.4.}
}

\value{a data frame of size pafdata containing the plotting 
positions of observations for each record.}
\references{Cunnane 1978, J. Hydrol. 37:205-222
Helsel and Hirsch, Statistical Methods in Water Resources, USGS.}
\author{Michael Koohafkan}




\examples{
data(esptraces)
esp.ahoc = subset_forecast(esptraces, location='AHOC1')
peakflows = get_peakavg(esp.ahoc)
get_ppositions(peakflows, 0.4) # Cunnane (default)
get_ppositions(peakflows, 0) # Weibull
}
