\name{plot_curves}
\alias{plot_curves}
\title{Plot empirical and fitted exceedance curves.}
\description{plot the exceedance curves and empirical data.}
\usage{plot_curves(flowdata = NULL, ppos = NULL, curvedata = NULL, distr = NULL, 
    units = c("af", "cf"), logscale = TRUE)}
\arguments{
  \item{flowdata}{flow data generated by get_peakavg().}
  \item{ppos}{plotting positions generated by get_ppositions().}
  \item{curvedata}{fitted curves generated by get_excurves().}
  \item{distr}{the distribution used to generate the fitted curves.}
  \item{units}{the units to be used for displaying flow. Options are cubic feet ('cf')
or acre-feet ('af').}
  \item{logscale}{logical: display log 10 scale on y-axis?}
}

\value{a ggplot2 plot.}

\author{Michael Koohafkan}




\examples{
data(workedexample)
plot_curves(usgs.fd, usgs.pp, usgs.fit, distr='lp3', logscale=TRUE)
plot_curves(usgs.fd, usgs.pp)
plot_curves(curvedata=usgs.fit, distr='lp3')
}
