\name{subset_forecast}
\alias{subset_forecast}
\title{Subset ESP trace data}
\description{pull time interval and/or specific locations from forecast data.}
\usage{subset_forecast(d, sc = c("GMT", "WYD"), startdate = NULL, enddate = NULL, 
    location = NULL)}
\arguments{
  \item{d}{a dataframe generated by get_forecast.}
  \item{sc}{name or index of column to use for date, e.g. 'GMT' or 'WYD'. 
Default is 'GMT'.}
  \item{startdate}{the start of interval, date format e.g. as.Date('2012-01-15')
or string 'YYYY-MM-DD', or numeric water year date, 
e.g. October 1st, 2013 is 2014001.}
  \item{enddate}{the last entry in interval, same format as startdate.}
  \item{location}{the region identifier, e.g. Arroyo Hondo is 'AHOC1'.}
}

\value{return subset of data.}

\author{Michael Koohafkan}




\examples{
data(esptraces)
subset_forecast(esptraces, location='AHOC1')
subset_forecast(esptraces, sc='GMT', startdate=Sys.Date()+7, enddate=Sys.Date()+183)
subset_forecast(esptraces, sc='WYD', enddate=to_wateryear(Sys.Date()+183))
}
